fix: Changed the message of the readme file
I had to change the message of the README.md file
because I wasn't really in my full brain power when
I wrote the last one.

I wrote in a drowsy state.

The reason why it's important to do all that they
instructed is so I don't get lost along the way and
I actually discipline myself to follow the actual
process.

@TOP @SOLOPROGRAMMING @PRACTICEMAKESPERFECT @GRIND