feat: Created the pages for the recipes website
I created the index.html file as well as the
recipes folder which houses all the other pages for
the recipes. I also created the images folder and
added the images to be used in the folder.

I previously used a forward slash to point to the
root in order to easily navigate to the index.html
but it caused an issue in the browser so i used
the ../ instead.